# openapi.yaml
openapi: 3.0.3
info:
  title: Billr API
  description: |
    B2B subscription billing API. Manage accounts, plans, subscriptions, usage metering,
    invoices, and webhooks. All monetary values are integers in the smallest currency unit
    (e.g. cents for USD). Never floats.

    ## Authentication
    All endpoints require a Bearer API key in the `Authorization` header, except
    `GET /plans` and `GET /plans/{planId}` which are public.

    ## Rate Limiting
    Requests are rate-limited per API key. Limits are returned in response headers:
    - `X-RateLimit-Limit` — requests allowed per window
    - `X-RateLimit-Remaining` — requests remaining
    - `X-RateLimit-Reset` — UTC epoch seconds when window resets

    ## Pagination
    List endpoints use cursor-based pagination. Pass `after` (a record ID) and `limit`
    (default 20, max 100). If `next_cursor` is present in the response there are more pages.

    ## Idempotency
    POST endpoints that create resources support an `Idempotency-Key` request header.
    Supply a unique UUID per logical operation; duplicate keys within 24 h return the
    cached response without re-executing the operation.
  version: 1.0.0
  contact:
    name: Billr Support
    email: support@billr.io
    url: https://docs.billr.io
  license:
    name: Proprietary
servers:
  - url: https://api.billr.io/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /accounts:
    post:
      tags: [Accounts]
      operationId: createAccount
      summary: Create account
      description: Creates a new customer account. `email` is used for auth; `billing_email` receives invoices (defaults to `email` if omitted).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
            example:
              name: "Acme Corp"
              email: "admin@acme.com"
              billing_email: "billing@acme.com"
              currency: USD
              timezone: "America/New_York"
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /accounts/{accountId}:
    get:
      tags: [Accounts]
      operationId: getAccount
      summary: Get account
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Account object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags: [Accounts]
      operationId: updateAccount
      summary: Update account
      description: Partial update — only provided fields are modified.
      parameters:
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountRequest'
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Accounts]
      operationId: deleteAccount
      summary: Delete account (hard)
      description: Immediately deletes the account and all associated data. Prefer `/accounts/{accountId}/delete` for GDPR erasure flows which queue an async deletion with audit trail.
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '204':
          description: Account deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /accounts/{accountId}/data-export:
    post:
      tags: [Accounts]
      operationId: requestDataExport
      summary: Request GDPR data export
      description: |
        Queues a full data export for the account (GDPR Article 20 — data portability).
        The export is emailed to `billing_email` within 24 hours as a JSON archive.
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '202':
          description: Export queued
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
                    example: "Export queued, you will receive an email within 24 hours"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /accounts/{accountId}/delete:
    post:
      tags: [Accounts]
      operationId: requestAccountDeletion
      summary: Request GDPR erasure (async)
      description: |
        Initiates an asynchronous account deletion (GDPR Article 17 — right to erasure).
        Data is scrubbed within 30 days. An audit record is retained as required by law.
        Active subscriptions are cancelled immediately.
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '202':
          description: Deletion initiated
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
                    example: "Account deletion initiated"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ---------------------------------------------------------------------------
  # Plans
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from POST /auth/keys. Prefix with `Bearer `.
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
      description: Account ID (prefix `acc_`)
...
  responses:
    BadRequest:
      description: Invalid request body or query parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: validation_error
            message: "'billing_interval' must be one of: monthly, yearly"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: unauthorized
            message: "Invalid or missing Bearer token"
    Forbidden:
      description: API key does not have permission for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: forbidden
            message: "read_only keys cannot perform this action"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: not_found
            message: "Account acc_01HX3Y9ZG2VMBPN5Q8T7RWKD not found"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: rate_limited
            message: "Rate limit exceeded. Retry after 30 seconds."
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: internal_error
            message: "An unexpected error occurred. Our team has been notified."
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: not_found
        message:
          type: string
          description: Human-readable description
          example: "Account not found"

    # -------------------------------------------------------------------------
    # Auth
    # -------------------------------------------------------------------------
...
    Account:
      type: object
      required: [id, name, email, currency, timezone, created_at, updated_at]
      properties:
        id:
          type: string
          example: acc_01HX3Y9ZG2VMBPN5Q8T7RWKD
        name:
          type: string
          example: Acme Corp
        email:
          type: string
          format: email
          example: admin@acme.com
        billing_email:
          type: string
          format: email
          nullable: true
          example: billing@acme.com
        currency:
          type: string
          description: ISO 4217 currency code
          example: USD
        timezone:
          type: string
          description: IANA timezone name
          example: America/New_York
        stripe_customer_id:
          type: string
          nullable: true
          description: Stripe customer ID — set after first checkout session
          example: cus_NffrFeUfNV2Hib
        metadata:
          type: object
          additionalProperties:
            type: string
          nullable: true
          description: Arbitrary key-value pairs for your own use
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAccountRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
        billing_email:
          type: string
          format: email
        currency:
          type: string
          default: USD
          description: ISO 4217 code
        timezone:
          type: string
          default: UTC
          description: IANA timezone name
        metadata:
          type: object
          additionalProperties:
            type: string
    UpdateAccountRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        billing_email:
          type: string
          format: email
        metadata:
          type: object
          additionalProperties:
            type: string

    # -------------------------------------------------------------------------
    # Plans
    # -------------------------------------------------------------------------
...